some changes

+++ runall.sh
@@ -8,6 +8,7 @@ sudo sh get-docker.sh
 sudo usermod -aG docker ubuntu
 
 # aws
+# awscli is not required but install it because it is usefull anyway.
 sudo apt install awscli -y
 mkdir ~/.aws
 touch ~/.aws/credentials
@@ -18,13 +19,17 @@ WRC_USERNAME='xxx' WRC_PASSWORD='xxx' kit='IRIS-2020.1.0.215.0-lnxubuntux64.tar.
 mv *.gz ~/
 sudo docker load -i ~/icm-2020.1.0.215.0-docker.tar.gz
 
-# download key
-# s3? need a s3 role.
-
 # for icm-utils
 sudo apt install jq -y
 cd ~
 git clone https://github.com/IRISMeister/icm-util.git
 cd icm-util
-cp ~/iris.key ./irislicense
 mv ~/IRIS-2020.1.0.215.0-lnxubuntux64.tar.gz ./kits
+
+# download iris.key from somewrhere...s3? 
+#
+if [ -e ~/iris.key ]; then
+  cp ~/iris.key ./irislicense
+else
+  echo "License key (iris.key) doesn't exist!"
+fi
